Finalize the support of no-suid with sycontainerize. Merge the two packages related to Singularity itself.
