+++ internal/pkg/network/network_ib.go
@@ -12,7 +12,7 @@ import (
 	"strconv"
 
 	"github.com/sylabs/singularity-mpi/internal/pkg/kv"
-	"github.com/sylabs/singularity-mpi/internal/pkg/util/sy"
+	"github.com/sylabs/singularity-mpi/internal/pkg/sy"
 	"github.com/sylabs/singularity-mpi/internal/pkg/sys"
 )
 
@@ -21,7 +21,7 @@ const (
 	IBForceKey = "force_ib"
 
 	// MXMDirKey is the key used in the configuration file to specify where MXM files are installed
-	MXMDirKey  = "mxm_dir"
+	MXMDirKey = "mxm_dir"
 
 	// KNEMDirKey is the key used in the configuration file to specify where knem files are installed
 	KNEMDirKey = "knem_dir"
